Prevent overflow when parsing numbers
at least make it happen with much bigger numbers only.
(cherry picked from commit ce7c7da7500ecbefb5090f262329ed9e9f220c23)

+++ src/common/calculator.c
@@ -62,7 +62,7 @@ typedef struct parser_state_t
 
 static float read_number(parser_state_t *self)
 {
-  int integer = 0, fractional = 0, fractional_div = 1;
+  unsigned long int integer = 0, fractional = 0, fractional_div = 1;
   while(*self->p && *self->p >= '0' && *self->p <= '9')
   {
     integer = integer * 10 + (*self->p) - '0';
